implemented multiple select of tags in composer

+++ src/main/kotlin/br/com/dblogic/blogkotlin/controller/TagController.kt
@@ -28,8 +28,8 @@ class TagController {
     }
 
     @GetMapping("/delete")
-    fun delete(@RequestParam(defaultValue = "0") id: Int){
-        if(id != 0) {
+    fun delete(@RequestParam(defaultValue = "0") id: Long){
+        if(id != 0L) {
             tagService.delete(Tag(id))
         }
     }
