+++ src/main/kotlin/br/com/dblogic/blogkotlin/model/facade/TagFacade.kt
@@ -1,6 +1,6 @@
 package br.com.dblogic.blogkotlin.model.facade
 
-data class TagFacade(val id: Int = 0,
+data class TagFacade(val id: Long = 0,
                      var name: String = "", 
                      val selected: Boolean = false) {
 }
